Allow multiple names to be passed at once + add support for api_key
An array of names can now be passed to Recognizer(), and an array of
results will be returned. For backwards-compatibility, if a single
string (non-array) is passed, a non-array is returned (as it functions
today).

This commit also adds support for an api_key to be passed (generated at
store.genderizer.io) via the constructor or
$recognizer->set_api_key($api_key).